refactor(useCounter): 물건 수량 조절하는 비즈니스 로직 커스텀 훅으로 분리
