+++ src/components/CartItem/CountButtonContainer.tsx
@@ -1,53 +1,17 @@
 import { css } from '@emotion/react';
-import { useSetRecoilState } from 'recoil';
 
-import { updateItemQuantity } from '@/apis/cartItem';
-import { MINUS, PLUS } from '@/assets/images';
-import { THEME } from '@/constants/theme';
-import { cartItemsState } from '@/recoil/cartItems/atoms';
 import { CartItemProps } from '@/types/cartItem';
+import { MINUS, PLUS } from '@assets/images';
+import { THEME } from '@constants/theme';
+
+import useCounter from '@hooks/useCounter';
 
 interface CountButtonContainerProps {
   item: CartItemProps;
 }
 
 const CountButtonContainer = ({ item }: CountButtonContainerProps) => {
-  const setCartItems = useSetRecoilState(cartItemsState);
-
-  const handleDecrementQuantity = async () => {
-    try {
-      const newQuantity = Math.max(item.quantity - 1, 1);
-      const { status } = await updateItemQuantity(item.id, newQuantity);
-
-      if (status === 200) {
-        setCartItems((prevItems) =>
-          prevItems.map((cartItem) =>
-            cartItem.id === item.id ? { ...cartItem, quantity: newQuantity } : cartItem,
-          ),
-        );
-      }
-    } catch (err: unknown) {
-      const error = err as Error;
-      throw new Error(error.message);
-    }
-  };
-
-  const handleIncrementQuantity = async () => {
-    try {
-      const { status } = await updateItemQuantity(item.id, item.quantity + 1);
-
-      if (status === 200) {
-        setCartItems((prevItems) =>
-          prevItems.map((cartItem) =>
-            cartItem.id === item.id ? { ...cartItem, quantity: cartItem.quantity + 1 } : cartItem,
-          ),
-        );
-      }
-    } catch (err: unknown) {
-      const error = err as Error;
-      throw new Error(error.message);
-    }
-  };
+  const { handleDecrementQuantity, handleIncrementQuantity } = useCounter(item);
 
   return (
     <div css={countWrapper}>
